changed Cycles standard distance from wall don't remove original cycle from cycle scene anymore but move it out of viewable space instead changed type of MAP_SIZE from int to ushort changed type of Round id from int to byte remove Cycles and Walls in function nullVars() which is called at the end of a round

+++ Core/GameSettings.cs
@@ -7,7 +7,7 @@ namespace Fusee.FuFiCycles.Core {
 		public static List<Round> ROUNDS = new List<Round>();
 		public static bool SHOW_MINIMAP = true;
 		public static byte PLAYER_QUANTITY = 2;
-		public static int MAP_SIZE;
+		public static ushort MAP_SIZE; // max size 65535
 		public static float4x4 SCENE_SCALE = float4x4.CreateScale(0.04f);
 	}
 }
